Grafic/requirements_gui.txt
PySide6>=6.5.0
torch>=2.9.1
torchvision>=0.24.1
einops>=0.8.1
Pillow>=11.3.0
pandas>=2.3.3
numpy>=2.3.3
matplotlib>=3.10.7
opencv-python>=4.5.0